refector mqtt app to use jinja to render templates, break down large code file into individual files for a cleaner read
