+++ mqtt/apps/MQTT Data Monitor/static/css/style.css
@@ -1,0 +1,214 @@
+/* MQTT Theme Status Colors */
+.status-new { 
+    background-color: #e3f2fd; 
+    color: #0078D4; 
+}
+
+.status-in_progress { 
+    background-color: #fff3e0; 
+    color: #ff8c00; 
+}
+
+.status-resolved { 
+    background-color: #e8f5e8; 
+    color: #00a86b; 
+}
+
+/* MQTT Theme Priority Colors */
+.severity-critical { 
+    background-color: #ffebee; 
+    color: #e74c3c; 
+}
+
+.severity-warning { 
+    background-color: #fff3e0; 
+    color: #ff8c00; 
+}
+
+.severity-info { 
+    background-color: #e3f2fd; 
+    color: #0078D4; 
+}
+
+/* MQTT Hover Effects */
+.mqtt-hover:hover { 
+    background-color: #f8f9fa; 
+    transition: all 0.2s ease; 
+}
+
+/* Tab Styles */
+.tab-button.active {
+    border-color: #0078D4 !important;
+    color: #0078D4 !important;
+}
+
+.tab-content.hidden {
+    display: none;
+}
+
+/* Custom Responsive Styles */
+@media (max-width: 768px) {
+    .grid-cols-1 {
+        grid-template-columns: repeat(1, minmax(0, 1fr));
+    }
+    
+    .md\:grid-cols-2 {
+        grid-template-columns: repeat(1, minmax(0, 1fr));
+    }
+    
+    .lg\:grid-cols-4 {
+        grid-template-columns: repeat(1, minmax(0, 1fr));
+    }
+    
+    .space-x-8 > :not([hidden]) ~ :not([hidden]) {
+        margin-left: 1rem;
+    }
+    
+    .px-4 {
+        padding-left: 0.5rem;
+        padding-right: 0.5rem;
+    }
+}
+
+/* Loading Animation */
+.loading {
+    position: relative;
+}
+
+.loading::after {
+    content: '';
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    width: 20px;
+    height: 20px;
+    margin: -10px 0 0 -10px;
+    border: 2px solid #f3f3f3;
+    border-top: 2px solid #0078D4;
+    border-radius: 50%;
+    animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+    0% { transform: rotate(0deg); }
+    100% { transform: rotate(360deg); }
+}
+
+/* Custom Form Elements */
+.form-input:focus {
+    outline: none;
+    border-color: #0078D4;
+    box-shadow: 0 0 0 3px rgba(0, 120, 212, 0.1);
+}
+
+/* Notification Styles */
+.notification-enter {
+    animation: slideIn 0.3s ease-out;
+}
+
+.notification-leave {
+    animation: slideOut 0.3s ease-in;
+}
+
+@keyframes slideIn {
+    from {
+        transform: translateY(-100%);
+        opacity: 0;
+    }
+    to {
+        transform: translateY(0);
+        opacity: 1;
+    }
+}
+
+@keyframes slideOut {
+    from {
+        transform: translateY(0);
+        opacity: 1;
+    }
+    to {
+        transform: translateY(-100%);
+        opacity: 0;
+    }
+}
+
+/* Table Enhancements */
+.table-row-hover:hover {
+    background-color: rgba(0, 120, 212, 0.05);
+    transition: background-color 0.15s ease;
+}
+
+/* Button Enhancements */
+.btn-primary {
+    background-color: #0078D4;
+    transition: all 0.2s ease;
+}
+
+.btn-primary:hover {
+    background-color: #106ebe;
+    transform: translateY(-1px);
+    box-shadow: 0 4px 8px rgba(0, 120, 212, 0.3);
+}
+
+.btn-secondary {
+    background-color: #037da5;
+    transition: all 0.2s ease;
+}
+
+.btn-secondary:hover {
+    background-color: #026d91;
+    transform: translateY(-1px);
+    box-shadow: 0 4px 8px rgba(3, 125, 165, 0.3);
+}
+
+/* Card Enhancements */
+.card-shadow {
+    box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06);
+    transition: box-shadow 0.3s ease;
+}
+
+.card-shadow:hover {
+    box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05);
+}
+
+/* Search Input Enhancements */
+.search-input {
+    transition: all 0.2s ease;
+}
+
+.search-input:focus {
+    transform: scale(1.02);
+    box-shadow: 0 4px 12px rgba(0, 120, 212, 0.15);
+}
+
+/* Badge Styles */
+.badge {
+    display: inline-flex;
+    align-items: center;
+    padding: 0.25rem 0.75rem;
+    border-radius: 9999px;
+    font-size: 0.75rem;
+    font-weight: 500;
+    text-transform: uppercase;
+    letter-spacing: 0.025em;
+}
+
+.badge-success {
+    background-color: rgba(0, 168, 107, 0.1);
+    color: #00a86b;
+}
+
+.badge-error {
+    background-color: rgba(231, 76, 60, 0.1);
+    color: #e74c3c;
+}
+
+.badge-warning {
+    background-color: rgba(255, 140, 0, 0.1);
+    color: #ff8c00;
+}
+
+.badge-info {
+    background-color: rgba(0, 120, 212, 0.1);
+    color: #0078D4;
+}
